Got Table of Contents added on each page
Also included a script to serve the statically generated site.

Everything seems to be working very well.  Last thing will be to add interactive figures and
annotation folding.  But I found a couple of packages that might help quite a bit with that
part by allowing me to inject React widgets into the HTML that I'm rendering.