Dynamically loading search index
This eliminates a huge number of bytes from the statically generated pages.
The dynamic loading of the search index is done in a reasonably nice way
with the search input disabled while the index is being loaded and a clear
indication that it is loading.  I also added caching so hopefully visitors
won't really see it beyond the first page load.